requirements.txt
torch
torchvision
torchxrayvision
numpy
opencv-python
scikit-image
SimpleITK
grad-cam
ttach
pandas
matplotlib
seaborn
scikit-learn
tqdm
PyYAML
requests